projectedit is routable and tries to fill in form data
the projectedit page can be routed with a project
ID, and if so, tries to retrieve that project and load
its data into the input fields. from then on, it'll
hopefully be treated just like a new project--
either overwriting an existing project ID in the
database, or creating a new project.